backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.1
cryptography==45.0.4
defusedxml==0.7.1
dj-rest-auth==7.0.1
django==5.2.3
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pycparser==2.22
pyjwt==2.9.0
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
